+++ doc.go
@@ -4,6 +4,7 @@
 // See the sub-packages:
 //   - Arche wrapper & scheduler -- [github.com/mlange-42/arche-model/model]
 //   - General-purpose systems -- [github.com/mlange-42/arche-model/system]
-//   - Reporter systems -- [github.com/mlange-42/arche-model/reporter]
+//   - Reporter systems for data handling -- [github.com/mlange-42/arche-model/reporter]
+//   - Observers for data extraction -- [github.com/mlange-42/arche-model/observer]
 //   - Commonly used resources -- [github.com/mlange-42/arche-model/resource]
 package archemodel