+++ observer/doc.go
@@ -1,0 +1,4 @@
+// Package observer provides interfaces for observers that extract data from the ECS world.
+// Observers are required by reporter systems (see [github.com/mlange-42/arche-model/reporter]).
+// The observer extracts the data, while the reporter handles it by e.g. showing a plot or writing to a file.
+package observer