+++ reporter/csv.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/mlange-42/arche-model/model"
+	"github.com/mlange-42/arche-model/observer"
 	"github.com/mlange-42/arche/ecs"
 )
 
@@ -14,10 +14,10 @@ import (
 //
 // Writes one row to a CSV file per step.
 type CSV struct {
-	Observer       model.Observer // Observer to get data from.
-	File           string         // Path to the output file.
-	Sep            string         // Column separator. Default ",".
-	UpdateInterval int            // Update interval in model ticks.
+	Observer       observer.Row // Observer to get data from.
+	File           string       // Path to the output file.
+	Sep            string       // Column separator. Default ",".
+	UpdateInterval int          // Update interval in model ticks.
 	file           *os.File
 	header         []string
 	builder        strings.Builder
@@ -27,7 +27,7 @@ type CSV struct {
 // Initialize the system
 func (s *CSV) Initialize(w *ecs.World) {
 	s.Observer.Initialize(w)
-	s.header = s.Observer.Header(w)
+	s.header = s.Observer.Header()
 	if s.Sep == "" {
 		s.Sep = ","
 	}